Matrix based github action for docker hub (#19)
This PR creates a workflow which will first check how many files have
changed when pushing to main (specifically Dockerfiles) and creates a
matrix json which looks like this

```
{"include": [{"context": "dockerfiles/submitty/autograding_default/1", "tags": "submitty/autograding_default:1"}, ...]}
```
This json will have only those files which have changed in merging
commit. Then it will run all these json objects (in parallel i guess)
and will push any changed/newly added dockerfiles to the main branch.

The only caveat is that all the next added dockerfiles should follow a
folder structure when added
-Submitty
--DockerImages
---dockerfiles
----submitty
-----<docker_name/language>
------<tag/docker_tag>
-------Dockerfile

Also this workflow can only push dockerfiles in the submitty account so
the workflow will fail for dockerfiles that are present in the
submittyrpi folder.
Any suggestions are welcome on how this should be handled:
Option 1: Upload them to submitty account
Option 2: Create a new account for individual university (submittyrpi,
submittyusc, submittyucla) and handle it individually

I believe this PR is ready to be look at by others I have tested mostly
every edge case in my personal repo:
Edge cases like:
1) Empty commit - in this case the docker jobs wont run and the workflow
will not fail it will just skip them
2) Adding dockerfiles that are not already present in the dockerhub
account - in this case it works as expected and adds a new docker to
dockerhub

Any suggestions for edge cases are welcome.


EDIT: New changes
There is a new file from which we check the latest version of the
dockers. The workflow reads through this json file and updates only
those that require an update.