[Bugfix:Autograding] Fix misnamed Java alias (#33)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
/opt/openjdk11 is intended to be a link to the JDK installation folder,
however the destination name is wrong. When /opt/openjdk11 is added to
the path, it is unable to find the Java installation or run any java
commands.

### What is the new behavior?
This fixes the destination name in the alias command. The path should
now include the JDK and the image should successfully execute java
commands.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->